UI/requirements.txt
streamlit==1.36.0
torch
torchvision
transformers
accelerate
safetensors
pillow
matplotlib